refactor accumulator tests
